Bones piles can be ransacked by adjacent monsters
This is another EvilHack feature that ended up completely different from
the original. I didn't really like the idea of adding a global to store
the monster that killed you, or that the looting monster could grab
items that happened to be on the square anyway, or special
considerations made for early demon lord deaths. I ended up looking into
the bones-pile-generation code and figured out a way to do ransacking of
the player's inventory there, and found it simpler to allow any
adjacent monster to ransack the player's inventory (which I like more
anyway; if you've been surrounded and taken down by a group, why would
only the monster delivering the killing blow get to take your stuff?)