Prevent vanilla hell prefab from overwriting other ones
I believe this is a latent bug in vanilla: hellfill.lua is allowed to
place various hell prefabs anywhere on the map without checking to see
if they overlap with an existing prefab. In xNetHack, this is blocked
via not allowing the same halign to be chosen twice for the same level,
but the "column of water/lava" prefab introduced from vanilla did not
have this. This led to an impossible in which an aquatic monster
generated in a water spot in the Kraken prefab which then was
overwritten by normal floor in the column prefab, resulting in it hiding
out of water.

Fix this by, for now, limiting the column prefab to only generate 1, and
do it using rnd_halign so that it will not overlap with another prefab.